More size-quantities using "unsigned long int" instead of "unsigned int".
That should be safer.

+++ Satisfiability/Solvers/OKsolver/SAT2002/Einlesen.h
@@ -1,5 +1,5 @@
 // Oliver Kullmann, 4.2.2001 (Toronto)
-/* Copyright 2001 - 2007 Oliver Kullmann
+/* Copyright 2001 - 2007, 2023 Oliver Kullmann
 This file is part of the OKlibrary. OKlibrary is free software; you can redistribute
 it and/or modify it under the terms of the GNU General Public License as published by
 the Free Software Foundation and included in this library; either version 3 of the
@@ -26,7 +26,7 @@ License, or any later version. */
 
 typedef enum {Norm = 0, Sat = 1, Unsat = 2, Fehler = 3} TEIN;
 
-extern TEIN Einlesen(FILE *fp, unsigned int G);
+extern TEIN Einlesen(FILE *fp, unsigned long int G);
 
 extern void AufraeumenEinlesen(void);
